refactor: split amazon/client.ts into focused parser modules
